+++ src/main/java/racingcar/domain/car/property/Position.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car.property;
 
 public class Position {
